refactor: 이미지 URL 처리 로직 함수로 분리
