+++ src/components/cart/CartItem/CartItem.tsx
@@ -2,6 +2,7 @@ import useCartOperations from "../../../domains/cart/hooks/useCartOperations";
 import useCartToggle from "../../../domains/cart/hooks/useCartToggle";
 import { CartItemWithSelection } from "../../../domains/cart/types/response";
 import { formatCurrency } from "../../../utils/formatters";
+import { getImageUrl } from "../../../utils/imageUrl";
 import Checkbox from "../../@common/Checkbox/Checkbox";
 import QuantityCounter from "../../@shared/QuantityCounter/QuantityCounter";
 import * as S from "./CartItem.styles";
@@ -23,7 +24,7 @@ const CartItem = ({ item: { id, quantity, product, selected } }: Props) => {
         <S.DeleteButton onClick={() => deleteItem(id)}>삭제</S.DeleteButton>
       </S.CartItemHeader>
       <S.CartItemWrapper>
-        <S.CartItemImage $url={product.imageUrl} />
+        <S.CartItemImage $url={getImageUrl(product.imageUrl)} />
         <S.CartItemInfo>
           <S.CartItemName>{product.name}</S.CartItemName>
           <S.CartItemPrice>{formatCurrency(product.price)}</S.CartItemPrice>
