+++ src/components/order/OrderItem/OrderItem.styles.ts
@@ -1,13 +1,5 @@
 import styled from "@emotion/styled";
 
-const DEFAULT_IMAGE_URL = "./planet-default-image.svg";
-
-const isValidUrl = (url: string): boolean =>
-  Boolean(url && (url.startsWith("http://") || url.startsWith("https://")));
-
-export const getImageUrl = (url: string): string =>
-  isValidUrl(url) ? url : DEFAULT_IMAGE_URL;
-
 export const OrderItem = styled.div`
   width: 100%;
   padding-top: 12px;
@@ -21,7 +13,7 @@ export const OrderItem = styled.div`
 export const OrderItemImage = styled.div<{ $url: string }>`
   width: 112px;
   height: 112px;
-  background: no-repeat url(${({ $url }) => getImageUrl($url)});
+  background: no-repeat url(${({ $url }) => $url});
   background-size: cover;
   border-radius: 8px;
 `;
