+++ src/components/order/OrderItem/OrderItem.tsx
@@ -1,5 +1,6 @@
 import { CartItemWithSelection } from "../../../domains/cart/types/response";
 import { formatCurrency } from "../../../utils/formatters";
+import { getImageUrl } from "../../../utils/imageUrl";
 import * as S from "./OrderItem.styles";
 
 interface Props {
@@ -10,7 +11,7 @@ interface Props {
 const OrderItem = ({ quantity, product }: Props) => {
   return (
     <S.OrderItem>
-      <S.OrderItemImage $url={product.imageUrl} />
+      <S.OrderItemImage $url={getImageUrl(product.imageUrl)} />
       <S.OrderItemInfo>
         <S.OrderItemName>{product.name}</S.OrderItemName>
         <S.OrderItemPrice>{formatCurrency(product.price)}</S.OrderItemPrice>
